requirements.txt
streamlit>=1.30.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
chromadb>=0.4.22
requests>=2.31.0
beautifulsoup4>=4.12.0
